cleaned_project/backtesters/amm-rebalance-backtester/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pydantic>=2.0.0
matplotlib>=3.7.0
optuna>=3.0.0
pyyaml>=6.0
rich>=13.0.0
tqdm>=4.65.0
polars>=0.19.0
numba>=0.57.0
click>=8.1.0
python-dateutil>=2.8.0
pytz>=2023.3
seaborn>=0.12.0
scikit-learn>=1.3.0
